backend/requirements.txt
mediapipe>=0.10.11
numpy>=1.24
Pillow>=10.0
scipy>=1.11.4
pyttsx3>=2.90
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
trimesh>=4.4.3

# Optional (enable Coqui XTTS voice cloning)
TTS==0.22.0
torch==2.1.0
torchaudio==2.1.0

